+++ UpgradeWorld/UpgradeWorld.cs
@@ -10,7 +10,7 @@ public class UpgradeWorld : BaseUnityPlugin
 {
   const string GUID = "upgrade_world";
   const string NAME = "Upgrade World";
-  const string VERSION = "1.73";
+  const string VERSION = "1.74";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable